requirements.txt
streamlit
beautifulsoup4>=4.12.0
html2image>=2.1.1
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0